fix: Product interface type으로 이동
