+++ src/types/index.ts
@@ -24,3 +24,17 @@ export interface Seller {
   id: string;
   nickname: string;
 }
+export interface Product {
+  id: number;
+  title: string;
+  price: number;
+  status: string;
+  condition: string;
+  images: string[];
+  description: string;
+  createdAt: string;
+  category: Category;
+  seller: Seller;
+  likesCount: number;
+  isLiked: boolean;
+}
